Conflicts:
	regards-starters/multitenant-regards-autoconfigure/src/main/java/fr/cnes/regards/framework/multitenant/autoconfigure/tenant/LocalTenantResolver.java

+++ rs-microservice/regards-default-modules/regards-plugin-module/module-plugins-service/src/main/java/fr/cnes/regards/modules/plugins/service/PluginService.java
@@ -10,9 +10,9 @@ import java.util.Map;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
+import fr.cnes.regards.framework.module.rest.exception.ModuleEntityNotFoundException;
 import fr.cnes.regards.modules.plugins.annotations.Plugin;
 import fr.cnes.regards.modules.plugins.dao.IPluginConfigurationRepository;
 import fr.cnes.regards.modules.plugins.domain.PluginConfiguration;
@@ -27,14 +27,10 @@ import fr.cnes.regards.plugins.utils.PluginUtilsException;
  * The implementation of {@link IPluginService}.
  *
  * @author Christophe Mertz
+ * @author Sébastien Binda
  */
 @Service
 public class PluginService implements IPluginService {
-
-    /**
-     * Package scanned to find plugins.
-     */
-    public static final String PLUGINS_PACKAGE = "fr.cnes.regards.plugins";
 
     /**
      * Class logger
@@ -42,9 +38,24 @@ public class PluginService implements IPluginService {
     private static final Logger LOGGER = LoggerFactory.getLogger(PluginService.class);
 
     /**
+     * The default Regards package for the {@link Plugin} of the core
+     */
+    private static final String REGARDS_PACKAGE_PLUGINS_DEFAULT = "fr.cnes.regards.plugins";
+
+    /**
+     * The default Regards package for the {@link Plugin} of the contributors
+     */
+    private static final String CONTRIB_PACKAGE_PLUGINS_DEFAULT = "fr.cnes.regards.contrib";
+
+    /**
+     * The plugin's package to scan
+     */
+    private final List<String> pluginPackage = new ArrayList<>();
+
+    /**
      * {@link PluginConfiguration} JPA Repository
      */
-    private IPluginConfigurationRepository pluginConfRepository;
+    private final IPluginConfigurationRepository pluginConfRepository;
 
     /**
      * Plugins implementation list sorted by plugin id. Plugin id, is the id of the "@PluginMetaData" annotation of the
@@ -54,37 +65,27 @@ public class PluginService implements IPluginService {
 
     /**
      * A constructor with the {@link IPluginConfigurationRepository}.
-     * 
+     *
      * @param pPluginConfigurationRepository
      *            {@link PluginConfiguration} JPA repository
-     * @throws PluginUtilsException
-     *             throw if an error occurs
      */
-    public PluginService(IPluginConfigurationRepository pPluginConfigurationRepository) throws PluginUtilsException {
+    public PluginService(final IPluginConfigurationRepository pPluginConfigurationRepository) {
         super();
         pluginConfRepository = pPluginConfigurationRepository;
-        this.loadPlugins();
+        getPluginPackage().add(REGARDS_PACKAGE_PLUGINS_DEFAULT);
+        getPluginPackage().add(CONTRIB_PACKAGE_PLUGINS_DEFAULT);
     }
 
-    /**
-     *
-     * Load from the classpath the plugins implementations of plugin annotation : {@link Plugin}
-     *
-     * @param parameters
-     *
-     * @throws PluginUtilsException
-     *             Error when loading the plugins
-     */
-    private void loadPlugins() throws PluginUtilsException {
-        // Scan class path for plugins implementations only once
+    private Map<String, PluginMetaData> getLoadedPlugins() {
         if (plugins == null) {
-            plugins = PluginUtils.getPlugins(PLUGINS_PACKAGE);
+            plugins = PluginUtils.getPlugins(getPluginPackage());
         }
+        return plugins;
     }
 
     @Override
     public List<String> getPluginTypes() {
-        return PluginInterfaceUtils.getInterfaces(PLUGINS_PACKAGE);
+        return PluginInterfaceUtils.getInterfaces(getPluginPackage());
     }
 
     @Override
@@ -96,10 +97,14 @@ public class PluginService implements IPluginService {
     public List<PluginMetaData> getPluginsByType(final Class<?> pInterfacePluginType) {
         final List<PluginMetaData> pluginAvailables = new ArrayList<>();
 
-        plugins.forEach((pKey, pValue) -> {
-            if (pInterfacePluginType == null || (pInterfacePluginType != null
-                    && pInterfacePluginType.isAssignableFrom(pValue.getPluginClass()))) {
-                pluginAvailables.add(pValue);
+        getLoadedPlugins().forEach((pKey, pValue) -> {
+            try {
+                if ((pInterfacePluginType == null) || ((pInterfacePluginType != null)
+                        && pInterfacePluginType.isAssignableFrom(Class.forName(pValue.getPluginClassName())))) {
+                    pluginAvailables.add(pValue);
+                }
+            } catch (final ClassNotFoundException e) {
+                LOGGER.error("cannot instanciate the class : %s" + pValue.getPluginClassName(), e);
             }
         });
 
@@ -110,29 +115,29 @@ public class PluginService implements IPluginService {
     public PluginConfiguration savePluginConfiguration(final PluginConfiguration pPluginConfiguration)
             throws PluginUtilsException {
         // Check plugin configuration validity
-        String message = "Impossible to save a plugin configuration";
+        final StringBuilder msg = new StringBuilder("Impossible to save a plugin configuration");
 
         boolean throwError = false;
 
         if (pPluginConfiguration == null) {
-            message += ". The plugin configuration cannot be null.";
+            msg.append(". The plugin configuration cannot be null.");
             throwError = true;
         }
-        if (!throwError && pPluginConfiguration.getPluginId() == null) {
-            message += ". The unique identifier of the plugin (attribute pluginId) is required.";
+        if (!throwError && (pPluginConfiguration.getPluginId() == null)) {
+            msg.append(". The unique identifier of the plugin (attribute pluginId) is required.");
             throwError = true;
         }
-        if (!throwError && pPluginConfiguration.getPriorityOrder() == null) {
-            message += String.format(" <%s> without priority order.", pPluginConfiguration.getPluginId());
+        if (!throwError && (pPluginConfiguration.getPriorityOrder() == null)) {
+            msg.append(String.format(" <%s> without priority order.", pPluginConfiguration.getPluginId()));
             throwError = true;
         }
-        if (!throwError && pPluginConfiguration.getVersion() == null) {
-            message = String.format(" <%s> without version.", pPluginConfiguration.getPluginId());
+        if (!throwError && (pPluginConfiguration.getVersion() == null)) {
+            msg.append(String.format(" <%s> without version.", pPluginConfiguration.getPluginId()));
             throwError = true;
         }
 
         if (throwError) {
-            throw new PluginUtilsException(message);
+            throw new PluginUtilsException(msg.toString());
         }
 
         return pluginConfRepository.save(pPluginConfiguration);
@@ -151,24 +156,21 @@ public class PluginService implements IPluginService {
 
     @Override
     public PluginConfiguration updatePluginConfiguration(final PluginConfiguration pPlugin)
-            throws PluginUtilsException {
+            throws ModuleEntityNotFoundException, PluginUtilsException {
         // Check if plugin configuration exists
         if (!pluginConfRepository.exists(pPlugin.getId())) {
-            throw new PluginUtilsException(
-                    String.format("Error while updating the plugin configuration <%s>.", pPlugin.getId()));
+            throw new ModuleEntityNotFoundException(pPlugin.getId().toString(), PluginConfiguration.class);
         }
-
         return savePluginConfiguration(pPlugin);
     }
 
     @Override
-    public void deletePluginConfiguration(final Long pPluginId) throws PluginUtilsException {
-        try {
-            pluginConfRepository.delete(pPluginId);
-        } catch (final EmptyResultDataAccessException e) {
-            throw new PluginUtilsException(
-                    String.format("Error while deleting the plugin configuration <%s>.", pPluginId), e);
+    public void deletePluginConfiguration(final Long pPluginId) throws ModuleEntityNotFoundException {
+        if (!pluginConfRepository.exists(pPluginId)) {
+            LOGGER.error(String.format("Error while deleting the plugin configuration <%s>.", pPluginId));
+            throw new ModuleEntityNotFoundException(pPluginId.toString(), PluginConfiguration.class);
         }
+        pluginConfRepository.delete(pPluginId);
     }
 
     @Override
@@ -177,20 +179,20 @@ public class PluginService implements IPluginService {
 
         final List<PluginMetaData> pluginImpls = getPluginsByType(pInterfacePluginType);
 
-        pluginImpls.forEach(p -> configurations
-                .addAll(pluginConfRepository.findByPluginIdOrderByPriorityOrderDesc(p.getPluginId())));
+        pluginImpls.forEach(pMetaData -> configurations
+                .addAll(pluginConfRepository.findByPluginIdOrderByPriorityOrderDesc(pMetaData.getPluginId())));
 
         return configurations;
     }
 
     @Override
-    public List<PluginConfiguration> getPluginConfigurationsByType(String pPluginId) {
+    public List<PluginConfiguration> getPluginConfigurationsByType(final String pPluginId) {
         return pluginConfRepository.findByPluginIdOrderByPriorityOrderDesc(pPluginId);
     }
 
     @Override
     public PluginMetaData getPluginMetaDataById(final String pPluginImplId) {
-        return plugins.get(pPluginImplId);
+        return getLoadedPlugins().get(pPluginImplId);
     }
 
     @Override
@@ -232,7 +234,7 @@ public class PluginService implements IPluginService {
         final PluginConfiguration pluginConf = getPluginConfiguration(pPluginConfigurationId);
 
         // Get the plugin implementation associated
-        final PluginMetaData pluginMetadata = plugins.get(pluginConf.getPluginId());
+        final PluginMetaData pluginMetadata = getLoadedPlugins().get(pluginConf.getPluginId());
 
         // Check if plugin version has changed since the last saved configuration of the plugin
         if ((pluginConf.getVersion() != null) && !pluginConf.getVersion().equals(pluginMetadata.getVersion())) {
@@ -241,6 +243,15 @@ public class PluginService implements IPluginService {
         }
 
         return PluginUtils.getPlugin(pluginConf, pluginMetadata, pPluginParameters);
+    }
+
+    public List<String> getPluginPackage() {
+        return pluginPackage;
+    }
+
+    @Override
+    public void addPluginPackage(final String pPluginPackage) {
+        this.pluginPackage.add(pPluginPackage);
     }
 
 }
